Add SSH setup, delay completions
SSH:
Add a doc and SSH config file for my LAN setup. This is going to be
different for people who decide to use this profile besides me, but
explains my thought process in a way that should be reproduceable.

Zsh completions:
I was finding that zsh was eating the letter 'i' in `git` if I wasn't
typing slowly because it was blocked waiting for a completion. This is
very annoying. Wait for 2 letters before starting completions to improve
responsiveness.